requirements.txt
flask
flask-cors
flask-security-too
flask-sqlalchemy
redis
pandas
celery